// src/domains/palworld-schema.ts
export enum DeathPenalty {
  None = "None",
  Item = "Item",
  ItemAndEquipment = "ItemAndEquipment",
  All = "All",
}
export enum AllowConnectPlatform {
  Steam = "Steam",
  Xbox = "Xbox",
}
export enum LogFormatType {
  Text = "Text",
  Json = "Json",
}